Completed the initial telemtery collector services and tested the api for telemetry event, have also added the cross cutting concern to builidng block that will be reused by all the microservices,

+++ src/Services/TelemetryCollector/TelemetryCollector.Domain/Models/TelemetryEvent.cs
@@ -6,7 +6,7 @@
          string EndPoint,   
          string Method,
          int StatusCode,
-         int ResponseTimeMs,
+         long ResponseTimeMs,
          DateTime Timestamp,
          string CorrelationId
     )
